+++ documentation/docs/changelog.md
@@ -6,6 +6,9 @@ sidebar_position: 2
 
 ## yyyy.mm.dd - version 1.13.0
 
+* Add [Settings](resources/core/settings) to handle player settings
+* Add Str:ucFirst() to capitalize the first letter of a string
+
 ## 2024.08.23 - version 1.12.2
 
 * Update wrong LuaDoc tags in some classes to fix the technical documentation